server: resolve concurrent errors within the server pkg
